implement quick-set encoder stops
i.e., when you want to quickly change the exposure time to the next
stop up or down, a single click of the encoder knob is all it takes.
you don't need to press SET unless you want to set a specific MM:SS
value (and you can use the encoder knob for that too, with each
value moving up/down with each click of the knob).

also make an attempt at debouncing the encoder by not reporting any
changes within 10ms of a reported click