terrible hack to reduce visual glitches at low brightness
at low brightness, the interrupt that turns the LEDs off occurs
very soon after the one that turned them on, and if the processor
is busy handling pin-change interrupts from the encoder, it can
be missed entirely. this results in a digit briefly appearing
at maximum brightness, exactly the kind of visual sparkle you
would not want to see when operating this thing at minimum
brightness.

... so set another compare-match interrupt to turn the LEDs off
a bit later. then at least the sparkles will be at the second-
lowest brightness level instead of the highest