fix display glitches frd
the blanking interrupt was not "missed" as the last commit
surmised. it just ran in the wrong order relative to
the refresh interrupt.

the problem occurred when both the overflow and compare
match A interrupt flags were set before either vector ran
(because the CPU was busy paying attention to a pin change
interrupt or whatever).

this commit ensures the refresh interrupt vector has a
higher priority than the blanking interrupt