tweak brightness table
the timer priority adjustment that got rid of the bright sparkles
in the lowest brightness caused the opposite problem at the
highest brightness level. fortunately, there is very little
visual difference between a 90% duty cycle and a 99% duty cycle.
so just give ourselves some time to separate the two events.