shut down after idle timeout
I'm pretty sure this thing was on in my camera bag all night,
because the power switch is easy to press accidentally.

fortunately the AVR is very good at powering itself down
and consumes a miniscule amount of power with just a pin-change
interrupt running to wake it up.

I measured a power consumption of like a fifth of a microamp
while this was shut down. if I flip the power switch off for a
few seconds and then back on, it's _still sleeping_, presumably
because the 100nf capacitor on the power rails has kept it
in that state!