make Set and Start buttons progress through EditNum
holding Set resets to 0, and Select increments by 1
(or 10 when held). of course using Select for this is
vestigial anyhow, because the encoder is much nicer.
I guess I'm leaving the capability to not use it,
in case I find it wigs out on me some night.

also fix a bug in display_spin() that I haven't actually witnessed
because my crystal doesn't take *that* long to start ticking