disable unused peripherals to reduce power consumption
the impact is small, unless the device is clocked at 8MHz,
when the savings become measurable. but since this would still
double power consumption at the lowest brightness level, I do
not find it compelling. we're staying at 1MHz for now,
but we'll make it easy to switch back if desired